Extract ProcessTracer into platform-specific files
Create a platform subdir and move the `ProcessTracer` class there. This
inherits from `AbstractProcessTracer` which provides a
platform-independent abstraction.

`ProcessManager` now requires a tracer to be provided. There have been
no uses of the fallback tracer code path, and this would complicate an
abstraction for multi-platform support as well, so we are now raising an
exception if a null tracer is provided.